+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Lifetime system : Add a lifetime component with a duration on an entity to make it disappear after the duration is elapsed
 - Physics System : Apply basic physics to rigidbodies and adds box colliders based collision detection
 - Virtual File System : Currently targetting a disk storage provider, but will be used later for packaging assets
+- Added a Asset Packer tool
+- Added a Package File provider for the VFS that reads Assets files packed with the dedicated tool
 
 ### Changed
 - Lifecycle is now a bit more precise with fixedUpdates and LateUpdates/RenderUpdates